feat(DSG): add LoginForm component to admin-ui static file and use it it the LogIn component
